move some methods to StringExpressionHandler
